Move image update methods back to separate file.
